✏️ Fix all the circular import error caused by the refactoring.
