+++ gearmeshing_ai/agent_core/capabilities/builtin.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Any, Dict
 
-from ..schemas.domain import CapabilityName
+from gearmeshing_ai.info_provider import CapabilityName
 from .base import Capability, CapabilityContext, CapabilityResult
