+++ test/e2e_test/agent_core/repos/test_sql_repos.py
@@ -31,12 +31,12 @@ from gearmeshing_ai.agent_core.schemas.domain import (
     AgentRun,
     Approval,
     ApprovalDecision,
-    CapabilityName,
     Checkpoint,
     RiskLevel,
     ToolInvocation,
     UsageLedgerEntry,
 )
+from gearmeshing_ai.info_provider import CapabilityName
 
 
 @pytest.fixture
