+++ test/e2e_test/agent_core/repos/test_sql_repositories_postgres.py
@@ -27,12 +27,12 @@ from gearmeshing_ai.agent_core.schemas.domain import (
     Approval,
     ApprovalDecision,
     AutonomyProfile,
-    CapabilityName,
     Checkpoint,
     RiskLevel,
     ToolInvocation,
     UsageLedgerEntry,
 )
+from gearmeshing_ai.info_provider import CapabilityName
 
 
 @pytest.fixture(scope="session")
