+++ test/e2e_test/agent_core/runtime/test_engine_integration.py
@@ -28,8 +28,8 @@ from gearmeshing_ai.agent_core.runtime.engine import AgentEngine
 from gearmeshing_ai.agent_core.schemas.domain import (
     AgentRun,
     AgentRunStatus,
-    CapabilityName,
 )
+from gearmeshing_ai.info_provider import CapabilityName
 
 
 class _OkCapability(Capability):
