+++ test/e2e_test/agent_core/runtime/test_thought_action_pause_resume.py
@@ -23,9 +23,9 @@ from gearmeshing_ai.agent_core.schemas.domain import (
     AgentRun,
     AgentRunStatus,
     ApprovalDecision,
-    CapabilityName,
     RiskLevel,
 )
+from gearmeshing_ai.info_provider import CapabilityName
 
 
 @pytest.mark.asyncio
