+++ test/e2e_test/agent_core/test_policy_provider.py
@@ -26,9 +26,9 @@ from gearmeshing_ai.agent_core.schemas.domain import (
     AgentRun,
     AgentRunStatus,
     AutonomyProfile,
-    CapabilityName,
     RiskLevel,
 )
+from gearmeshing_ai.info_provider import CapabilityName
 from gearmeshing_ai.agent_core.service import AgentService, AgentServiceDeps
