+++ test/integration_test/agent_core/test_policy_provider_engine_integration.py
@@ -23,9 +23,9 @@ from gearmeshing_ai.agent_core.schemas.domain import (
     AgentRun,
     AgentRunStatus,
     AutonomyProfile,
-    CapabilityName,
     RiskLevel,
 )
+from gearmeshing_ai.info_provider import CapabilityName
 from gearmeshing_ai.agent_core.service import AgentService, AgentServiceDeps
