+++ test/unit_test/agent_core/planning/test_planner.py
@@ -8,7 +8,7 @@ from gearmeshing_ai.agent_core.planning import (
     ActionStep,
     StructuredPlanner,
 )
-from gearmeshing_ai.agent_core.schemas.domain import CapabilityName
+from gearmeshing_ai.info_provider import CapabilityName
 
 
 @pytest.mark.asyncio
